requirements.txt
streamlit
tensorflow
numpy
pillow
streamlit-lottie
requests
matplotlib
pandas
opencv-python
scikit-learn